+++ _types/application/index.d.ts
@@ -1,4 +1,140 @@
 /**
+ * Provides the ability to mount and control Svelte component based sidebar panels & tabs in the Foundry sidebar.
+ *
+ * To add a new sidebar tab schedule one or more invocations of {@link FVTTSidebarControl.add} in a `setup` hook. You
+ * must add all sidebars in the `setup` hook before the main Foundry sidebar renders. Please review all the expanded
+ * options available in the configuration object passed to the `add` method. At minimum, you need to provide a unique
+ * `id`, `icon`, and `svelte` configuration object. You almost always will want to provide `beforeId` referencing
+ * another existing sidebar tab ID to place the tab button before. If undefined the tab is inserted at the end of
+ * the sidebar tabs.
+ *
+ * The nice thing about FVTTSidebarControl is that all you have to provide is the sidebar component and the rest is
+ * handled for you including automatically widening the width of the sidebar to fit the new sidebar tab. Also by default
+ * an adhoc SvelteApplication is configured to display the sidebar when popped out automatically without the need to
+ * associate an app instance.
+ *
+ * -------------------------------------------------------------------------------------------------------------------
+ *
+ * Optionally:
+ * - You can define the `icon` as a Svelte configuration object to load an interactive component instead of
+ * using a FontAwesome icon. This allows you to dynamically show state similar to the chat log sidebar when activity
+ * occurs or for other purposes.
+ *
+ * - You can provide `popoutOptions` overriding the default options passed to the default adhoc SvelteApplication
+ * rendered for the popout.
+ *
+ * - You can provide a class that extends from SvelteApplication as `popoutApplication` to provide a fully customized
+ * popout sidebar that you fully control.
+ *
+ * -------------------------------------------------------------------------------------------------------------------
+ *
+ * The {@link FVTTSidebarControl.get} method allows you to retrieve the associated {@link TJSSidebarEntry} for a given
+ * sidebar by ID allowing access to the configuration data, popout app, and wrapper components that mount the sidebar.
+ *
+ * The {@link FVTTSidebarControl.wait} returns a Promise that is resolved after all sidebars have been initialized.
+ * allowing handling any special setup as necessary.
+ *
+ * -------------------------------------------------------------------------------------------------------------------
+ *
+ * @example Minimal setup
+ *
+ * Hooks.once('setup', () =>
+ * {
+ *    FVTTSidebarControl.add({
+ *       beforeId: 'items',         // Place new tab before the 'items' tab
+ *       id: 'test',                // A unique CSS ID
+ *       icon: 'fas fa-dice-d10',   // FontAwesome icon
+ *       title: 'Test Directory',   // Title of popout sidebar app; can be language string.
+ *       tooltip: 'Tests',          // Tooltip for sidebar tab.
+ *       svelte: {
+ *          class: TestTab          // A Svelte component
+ *       }
+ *    });
+ * });
+ */
+declare class FVTTSidebarControl {
+    /**
+     * @type {object[]}
+     */
+    static "__#148349@#initData": object[];
+    static "__#148349@#initPromise": any;
+    /**
+     * @type {Map<string, TJSSidebarEntry>}
+     */
+    static "__#148349@#sidebars": Map<string, TJSSidebarEntry>;
+    /**
+     * @param {object}   sidebarData - The configuration object for a Svelte sidebar,
+     *
+     * @param {string}   sidebarData.id - The unique Sidebar ID / name. Used for CSS ID and retrieving the sidebar.
+     *
+     * @param {string|object}  sidebarData.icon - The FontAwesome icon css classes _or_ a Svelte configuration object
+     * to load a custom Svelte component to use as the "icon".
+     *
+     * @param {object}   sidebarData.svelte - A Svelte configuration object.
+     *
+     * @param {string}   [sidebarData.beforeId] - The ID for the tab to place the new sidebar before. This must be an
+     *        existing sidebar tab ID. THe stock Foundry sidebar tab IDs from left to right are:
+     *
+     * @param {string}   [sidebarData.popoutApplication] - Provides a custom SvelteApplication class to instantiate
+     *        for the popout sidebar.
+     *
+     * @param {string}   [sidebarData.popoutOptions] - Provides SvelteApplication options overrides for the default
+     *        popout sidebar.
+     *
+     * @param {string}   [sidebarData.title] - The popout application title text or i18n lang key.
+     *
+     * @param {string}   [sidebarData.tooltip] - The sidebar tab tooltip text or i18n lang key.
+     */
+    static add(sidebarData: {
+        id: string;
+        icon: string | object;
+        svelte: object;
+        beforeId?: string;
+        popoutApplication?: string;
+        popoutOptions?: string;
+        title?: string;
+        tooltip?: string;
+    }): void;
+    /**
+     * Initializes all sidebars registered after the initial Foundry Sidebar app has been rendered.
+     */
+    static "__#148349@#initialize"(): void;
+    /**
+     * Returns a loaded and configured sidebar entry by ID.
+     *
+     * @param {string}   id - The ID of the sidebar to retrieve.
+     *
+     * @returns {TJSSidebarEntry} The sidebar entry.
+     */
+    static get(id: string): TJSSidebarEntry;
+    /**
+     * Provides a Promise that is resolved after all added sidebars are initialized. This is useful when additional
+     * setup or configuration of sidebars needs to be performed after sidebar initialization.
+     *
+     * @returns {Promise} Initialization Promise.
+     */
+    static wait(): Promise<any>;
+}
+type TJSSidebarEntry = {
+    /**
+     * - The sidebar data that configures a Svelte sidebar.
+     */
+    data: object;
+    /**
+     * - The sidebar popout application.
+     */
+    popout: SvelteApplication;
+    /**
+     * - The tab wrapper component.
+     */
+    tab: FVTTSidebarTab;
+    /**
+     * - The sidebar wrapper component.
+     */
+    wrapper: FVTTSidebarWrapper;
+};
+
+/**
  * Provides and manages browser window wide context menu functionality. The best way to create a context menu is to
  * attach the source KeyboardEvent or MouseEvent / PointerEvent as data in {@link TJSContextMenu.create}. This allows
  * proper keyboard handling across browsers supporting the context menu key. A A11yFocusSource data object is generated
@@ -11,7 +147,7 @@ declare class TJSContextMenu {
     /**
      * Stores any active context menu.
      */
-    static "__#149718@#contextMenu": any;
+    static "__#148350@#contextMenu": any;
     /**
      * Creates and manages a browser wide context menu. The best way to create the context menu is to pass in the source
      * DOM event as it is processed for the location of the context menu to display. Likewise, a A11yFocusSource object
@@ -67,7 +203,7 @@ declare class TJSContextMenu {
      *
      * @returns {object[]} Processed menu items.
      */
-    static "__#149718@#processItems"(items: Iterable<TJSContextMenuItemData>): object[];
+    static "__#148350@#processItems"(items: Iterable<TJSContextMenuItemData>): object[];
 }
 /**
  * - Defines a menu item entry. Depending on the item data that is passed
@@ -116,4 +252,4 @@ type TJSContextMenuItemData = {
     separator?: 'hr';
 };
 
-export { TJSContextMenu, TJSContextMenuItemData };
+export { FVTTSidebarControl, TJSContextMenu, TJSContextMenuItemData, TJSSidebarEntry };