+++ src/component/fvtt/sidebar/FVTTSidebarTab.svelte
@@ -1,0 +1,40 @@
+<script>
+   /**
+    * INTERNAL USE ONLY: Provides the wrapper component for mounting a custom Svelte sidebar; used by
+    * `FVTTSidebarControl`.
+    */
+
+   /**
+    * Export the main anchor element so that FVTTSidebarControl can adjust the sidebar width.
+    *
+    * @type {HTMLAnchorElement}
+    */
+   export let anchorEl;
+
+   /**
+    * Bound sidebar tab component making it accessible from {@link TJSSidebarEntry}.
+    */
+   export let component = void 0;
+
+   /**
+    * Sidebar configuration data.
+    */
+   export let sidebarData = void 0;
+</script>
+
+<svelte:options accessors={true} />
+
+<!-- svelte-ignore a11y-missing-attribute -->
+<a bind:this={anchorEl}
+   class=item
+   data-tab={sidebarData.id}
+   data-tooltip={sidebarData.tooltip}
+   alt={sidebarData.tooltip}>
+   {#if sidebarData.iconSvelteConfig !== void 0}
+      <svelte:component bind:this={component}
+                        this={sidebarData.iconSvelteConfig.class}
+                        {...(sidebarData.iconSvelteConfig.props ?? {})} />
+   {:else}
+      <i class={sidebarData.icon}></i>
+   {/if}
+</a>