+++ src/component/fvtt/sidebar/index.js
@@ -1,0 +1,3 @@
+export { default as FVTTSidebarPopout }   from './FVTTSidebarPopout.svelte';
+export { default as FVTTSidebarTab }      from './FVTTSidebarTab.svelte';
+export { default as FVTTSidebarWrapper }  from './FVTTSidebarWrapper.svelte';